pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fit5226_project"
version = "0.1.0"
dependencies = [
    "numpy",
    "matplotlib",
    "torch",
    "scipy",
    "tqdm",
    "mlflow",
    "PyYAML",
    "fire",
    "optuna",
    "ipykernel",
    "optuna-dashboard",
]

[tool.ruff]
line-length = 119